test config renamed
